Factor out save recipe button to a separate component
