+++ frontend/src/components/recipes/RecipeCard.tsx
@@ -1,9 +1,8 @@
-import {Heart, Clock} from "lucide-react";
+import {Clock} from "lucide-react";
 import {Card, CardContent, CardHeader, CardTitle} from "@/components/ui/card";
 import {Button} from "@/components/ui/button";
 import {Recipe} from "@/hooks/queries/types.ts";
-import {useSaveRecipe, useUnsaveRecipe} from "@/hooks/mutations/useSaveRecipe.ts";
-import {useKeyboardShortcuts} from "@/hooks/useKeyboardShortcuts.ts";
+import {SaveRecipeButton} from "@/components/recipes/SaveRecipeButton.tsx";
 
 type Props = {
     recipe: Recipe;
@@ -18,14 +17,6 @@ type Props = {
  * */
 export const RecipeCard = (props: Props) => {
     const {recipe, onOpen, onBook, onHover, isFocused = false} = props;
-
-    const onSaveRecipe = useSaveRecipe(recipe.id);
-    const onUnsaveRecipe = useUnsaveRecipe(recipe.id);
-    const onToggleSaved = () => {
-        return recipe.isSaved ? onUnsaveRecipe.mutate() : onSaveRecipe.mutate();
-    }
-
-    useKeyboardShortcuts({'s': () => isFocused ? onToggleSaved() : undefined})
 
     return (
         <Card
@@ -62,12 +53,7 @@ export const RecipeCard = (props: Props) => {
                                     aria-label="Add to meal plan">
                                 Book
                             </Button>
-                            <Button variant={recipe.isSaved ? "secondary" : "outline"} size="sm"
-                                    onClick={onToggleSaved} aria-pressed={recipe.isSaved}
-                                    aria-label={recipe.isSaved ? "Unsave recipe" : "Save recipe"}>
-                                <Heart className={`h-4 w-4 ${recipe.isSaved ? "fill-current" : ""}`}/>
-                                {recipe.isSaved ? "Saved" : "Save"}
-                            </Button>
+                            <SaveRecipeButton recipe={recipe} isFocused={isFocused}/>
                         </div>
                     </div>
                 </div>
